% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powreg.R
\name{obj.varcomp}
\alias{obj.varcomp}
\title{Function for computing the profile likelihood of the variance ratio under normal-normal model}
\usage{
obj.varcomp(r.sq, y = y, X = X)
}
\arguments{
\item{\code{r.sq}}{variance ratio}

\item{\code{y}}{regression response}

\item{\code{X}}{regression design matrix}
}
\value{
Objective function value
}
\description{
\code{varcomp}
}
