\name{powreg-package}
\alias{powreg-package}
\alias{powreg}
\docType{package}
\title{
\packageTitle{powreg}
}
\description{
\packageDescription{powreg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{powreg}
\packageIndices{powreg}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{powreg}

Maintainer: \packageMaintainer{powreg}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
