% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powreg.R
\name{s.sq.r.sq}
\alias{s.sq.r.sq}
\title{Some functions for doing variance component estimation}
\usage{
s.sq.r.sq(r.sq, y.tilde = NULL, d = NULL, y = NULL, X = NULL)
}
\description{
Some functions for doing variance component estimation
}
