% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powreg.R
\name{varcomp}
\alias{varcomp}
\title{Function for computing maximum likelihood estimates of variance parameters under normal-normal model}
\usage{
varcomp(y, X)
}
\arguments{
\item{\code{y}}{regression response}

\item{\code{X}}{regression design matrix}
}
\value{
Estimates \code{sigma.beta.sq.hat}, \code{sigma.epsi.sq.hat}
}
\description{
\code{varcomp}
}
