% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powreg.R
\name{estRegPars}
\alias{estRegPars}
\title{Function for estimating tuning parameters}
\usage{
estRegPars(y, X, delta.sq = NULL, precomp = NULL, comp.q = FALSE,
  mom = TRUE)
}
\arguments{
\item{\code{y}}{regression response}

\item{\code{X}}{regression design matrix}

\item{\code{delta}}{ridge regression parameter for when X is not full rank}
}
\value{
Estimates \code{sigma.beta.sq.hat}, \code{sigma.epsi.sq.hat} and \code{kappa.hat}
}
\description{
\code{estRegPars}
}
